+++ com.github.eclipsecolortheme/src/com/github/eclipsecolortheme/Activator.java
@@ -9,6 +9,7 @@ public class Activator extends AbstractUIPlugin {
 	
     public static final String PLUGIN_ID = "com.github.eclipsecolortheme";
     public static final String EXTENSION_POINT_ID_MAPPER = PLUGIN_ID + ".mapper";
+    public static final String EXTENSION_POINT_ID_THEME = PLUGIN_ID + ".theme";
     
     private static Activator plugin;