+++ CHANGES.md
@@ -5,6 +5,8 @@
 * Cleaned up the Ant color keys.
 * Refactored mapping logic.
 * Added support for text decorations (bold, italic, underline, strikethrough).
+* Added an Extension Point for Mappers.
+* Added an Extension Point for Themes.
 
 0.8.0 (2011-02-07)
 ==================